stdio: Fix -Wtautological-constant-out-of-range-compare on clang
Clang emits an error:

error: result of comparison of constant 255 with expression of type
'char' is always true
[-Werror,-Wtautological-constant-out-of-range-compare]
          if (spec <= UCHAR_MAX

+++ stdio-common/vfprintf-internal.c
@@ -1380,7 +1380,7 @@ printf_positional (FILE *s, const CHAR_T *format, int readonly_format,
 	{
 	  int function_done;
 
-	  if (spec <= UCHAR_MAX
+	  if ((int) spec <= UCHAR_MAX
 	      && __printf_function_table != NULL
 	      && __printf_function_table[(size_t) spec] != NULL)
 	    {
